adiciona método para colorir botões
- método colorButtonsOnMouseEntered
- colore ao arrastar o mouse no botão
- adicionado na MainScreenView / utilizado por clear e addE

+++ src/grafando/Controller/ConnectVertexController.java
@@ -53,8 +53,8 @@ public class ConnectVertexController {
 
     private void callDelegateDrawEdges() {
         Integer initialVertex, finalVertex;
-        initialVertex = (Integer) view.getStartingVertex().getValue();
-        finalVertex = (Integer) view.getFinalVertex().getValue();
+        initialVertex = view.getStartingVertex().getValue();
+        finalVertex = view.getFinalVertex().getValue();
         mainScreenController.callDrawEdgeOnView(initialVertex,finalVertex);
     }
 }
